+++ lib/neighbor/model.rb
@@ -2,11 +2,9 @@ module Neighbor
   module Model
     def has_neighbors(*attribute_names, dimensions: nil, normalize: nil)
       if attribute_names.empty?
-        warn "[neighbor] has_neighbors without an attribute name is deprecated"
-        attribute_names << :neighbor_vector
-      else
-        attribute_names.map!(&:to_sym)
+        raise ArgumentError, "has_neighbors requires an attribute name"
       end
+      attribute_names.map!(&:to_sym)
 
       class_eval do
         @neighbor_attributes ||= {}
@@ -44,13 +42,7 @@ module Neighbor
           distance = options.delete(:distance)
           raise ArgumentError, "unknown keywords: #{options.keys.map(&:inspect).join(", ")}" if options.any?
 
-          if vector.nil? && !attribute_name.nil? && attribute_name.respond_to?(:to_a)
-            warn "[neighbor] nearest_neighbors without an attribute name is deprecated"
-            vector = attribute_name
-            attribute_name = :neighbor_vector
-          end
           attribute_name = attribute_name.to_sym
-
           options = neighbor_attributes[attribute_name]
           raise ArgumentError, "Invalid attribute" unless options
           normalize = options[:normalize]
@@ -139,11 +131,7 @@ module Neighbor
             .order(Arel.sql(order))
         }
 
-        def nearest_neighbors(attribute_name = nil, **options)
-          if attribute_name.nil?
-            warn "[neighbor] nearest_neighbors without an attribute name is deprecated"
-            attribute_name = :neighbor_vector
-          end
+        def nearest_neighbors(attribute_name, **options)
           attribute_name = attribute_name.to_sym
           # important! check if neighbor attribute before accessing
           raise ArgumentError, "Invalid attribute" unless self.class.neighbor_attributes[attribute_name]