+++ test/support/vector.rb
@@ -3,7 +3,6 @@ ActiveRecord::Schema.define do
 
   create_table :items, force: true do |t|
     t.vector :embedding, limit: 3
-    t.vector :neighbor_vector, limit: 3
     t.vector :factors, limit: 3
     t.halfvec :half_embedding, limit: 3
     t.bit :binary_embedding, limit: 3
@@ -12,7 +11,6 @@ end
 
 class Item < ActiveRecord::Base
   has_neighbors :embedding, :half_embedding, :binary_embedding
-  has_neighbors
 end
 
 class CosineItem < ActiveRecord::Base